+++ package.json
@@ -21,10 +21,12 @@
     "hbs": "^4.2.0",
     "ldapauth-fork": "^5.0.5",
     "ldapjs": "^3.0.7",
+    "moment": "^2.30.1",
     "mysql": "^2.18.1",
     "mysql2": "^3.6.5",
     "passport": "^0.7.0",
     "passport-ldapauth": "^3.0.1",
+    "speakeasy": "^2.0.0",
     "uuid": "^9.0.1"
   },
   "devDependencies": {
